+++ java-service-processor/src/test/resources/definition/TestIdUniqueness.java
@@ -1,0 +1,14 @@
+package definition;
+
+import nbbrd.service.ServiceDefinition;
+import nbbrd.service.ServiceId;
+
+@ServiceDefinition
+interface TestIdUniqueness {
+
+    @ServiceId
+    String getName();
+
+    @ServiceId
+    String getDescription();
+}